feat(frontend): resorts list show done

+++ Backend/FishingBooker/FishingBooker/src/main/java/com/isa/FishingBooker/service/interfaces/RevisionService.java
@@ -9,7 +9,7 @@ import com.isa.FishingBooker.model.revision.TutorServiceAppointmentRevision;
 public interface RevisionService extends Service<Revision> {
 	public void updateRevisionStatus(int revisionId, Status revisionStatus, int adminId);
 	public double getBoatAppointmentRevisionsRate(Integer id);
-	public int getResortAppointmentRevisionsRate(Integer id);
+	public double getResortAppointmentRevisionsRate(Integer id);
 	public double getTutorServiceAppointmentRevisionsRate(Integer id);
 	public String makeTutorServiceRevision(TutorServiceAppointmentRevision revision);
 	public String makeBoatRevision(BoatAppointmentRevision revision);
